Moving to integration tests
